requirements.txt
pydantic
passlib= 1.7.4
python-jose= 3.3.0
pytz= 2024.2
python-multipart= 0.0.9
